Reapply "Merge pull request #650 from alphagov/send_job_with_email_files"
This reverts commit e42b385db83bbf2eac80bf97b5807a23eeca83ef.

Test re-adding code against staging
