+++ tests/pages/pages.py
@@ -900,17 +900,10 @@ class ViewEmailTemplatePage(ViewTemplatePage):
         element = self.wait_for_element(ViewEmailTemplatePage.page_banner_text)
         return element.text.strip()
 
-    def click_delete_template_link(self):
-        element = self.wait_for_element(ViewEmailTemplatePage.delete_template_link)
-        element.click()
-
-    def click_template_deletion_confirmation_button(self):
-        element = self.wait_for_element(ViewEmailTemplatePage.template_deletion_confirmation_button)
-        element.click()
-
     def get_email_message_body_content(self):
         element = self.wait_for_element(ViewEmailTemplatePage.email_message_body_content)
         return element.text.strip()
+
     def click_file_link_text(self, link_text):
         element = self.wait_for_element((By.XPATH, f"//a[contains(text(), '{link_text}')]"))
         element.click()
@@ -1216,6 +1209,7 @@ class SendViaCsvPreviewPage(PageWithCsvPreview, PageWithSendToMultipleButton):
 class JobPage(BasePage):
     uploads_link = (By.LINK_TEXT, "Uploads")
     first_notification = JobPageLocators.FIRST_NOTIFICATION
+    notification_link = JobPageLocators.NOTIFICATION_LINK
 
     def wait_until_current(self, time=10):
         return self.wait_until_url_contains("/jobs/", time=time)
@@ -1229,6 +1223,14 @@ class JobPage(BasePage):
                 raise RetryException(f"No notification id yet {notification_id}")
             else:
                 return notification_id
+        except StaleElementReferenceException as e:
+            raise RetryException("Could not find element...") from e
+
+    @retry(RetryException, tries=20, delay=10)
+    def go_to_notification_page(self):
+        try:
+            notification_link = self.wait_for_element(self.notification_link)
+            notification_link.click()
         except StaleElementReferenceException as e:
             raise RetryException("Could not find element...") from e
